+++ src/main/java/gift/config/OpenApiConfig.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
