+++ src/main/java/gift/config/RestClientConfig.java
@@ -1,0 +1,25 @@
+package gift.config;
+
+import java.time.Duration;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+import org.springframework.web.client.RestClient;
+
+@Configuration
+public class RestClientConfig {
+
+    @Bean
+    public RestClient restClient() {
+        return RestClient.builder()
+            .requestFactory(timeout())
+            .build();
+    }
+
+    private HttpComponentsClientHttpRequestFactory timeout() {
+        var factory = new HttpComponentsClientHttpRequestFactory();
+        factory.setConnectTimeout(Duration.ofSeconds(5));
+        return factory;
+    }
+
+}
