+++ src/main/java/gift/order/OrderService.java
@@ -7,9 +7,7 @@ import gift.product.Product;
 import gift.wishes.Wish;
 import gift.wishes.WishService;
 import java.net.URI;
-import java.time.Duration;
 import org.springframework.http.MediaType;
-import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.web.client.RestClient;
@@ -20,13 +18,15 @@ public class OrderService {
     private final OrderRepository orderRepository;
     private final OptionService optionService;
     private final WishService wishService;
+    private final RestClient client;
 
 
     public OrderService(OrderRepository orderRepository,
-        OptionService optionService, WishService wishService) {
+        OptionService optionService, WishService wishService, RestClient client) {
         this.orderRepository = orderRepository;
         this.optionService = optionService;
         this.wishService = wishService;
+        this.client = client;
     }
 
     public OrderInfo saveOrder(OrderRequest orderRequest) {
@@ -55,9 +55,6 @@ public class OrderService {
     }
 
     public void sendMessage(OrderRequest orderRequest, String accessToken) {
-        RestClient client = RestClient.builder()
-            .requestFactory(timeout())
-            .build();
         String url = "https://kapi.kakao.com/v2/api/talk/memo/default/send";
 
         var template = setMessage(orderRequest);
@@ -76,12 +73,6 @@ public class OrderService {
         Product product = optionService.getProduct(optionId);
         Wish wish = wishService.getWish(product.getId(), memberId);
         wishService.deleteWish(wish.getId(), memberId);
-    }
-
-    private HttpComponentsClientHttpRequestFactory timeout() {
-        var factory = new HttpComponentsClientHttpRequestFactory();
-        factory.setConnectTimeout(Duration.ofSeconds(5));
-        return factory;
     }
